supply chain management
Supply Chain Management:
Supply chain management is the handling of the entire production flow of a good or service — starting from the raw components all the way to delivering the final product to the consumer. A company creates a network of suppliers that move the product along from the suppliers of raw materials to those organizations that deal directly with users.